+++ billing-payment/src/main/java/ma/atos/billing/payment/billing_payment/repositories/PaymentRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.jpa.repository.Query;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface PaymentRepository extends JpaRepository<Payment, Long>, JpaSpecificationExecutor<Payment> {
@@ -26,5 +27,5 @@ public interface PaymentRepository extends JpaRepository<Payment, Long>, JpaSpec
     long countByOperationType(String operationType);
 
     @Query("select coalesce(sum(p.amount), 0) from Payment p where p.status = 'SUCCESS'")
-    Double sumSuccessfulAmount();
+    BigDecimal sumSuccessfulAmount();
 }
